% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_SNV.R
\name{add_SNV}
\alias{add_SNV}
\title{Add a column with SNV data}
\usage{
add_SNV(df, gene_column = "SHORT_NAME")
}
\arguments{
\item{df}{A dataframe of imported snpSIFT data from mappgene output}
}
\description{
Add a column with SNV data
}
