% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zika_reference.R
\name{add_zika_reference}
\alias{add_zika_reference}
\title{Add data for processed Zika peptides}
\usage{
add_zika_reference(df, remove_seqs = TRUE)
}
\arguments{
\item{df}{A mappgene summary dataframe}

\item{remove_seqs}{Remove the nucleotide and amino acid sequences from the dataframe}
}
\description{
Given a mappgene dataframe, this function will add the genome reference data
from Zika KJ776791.
}
