% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_summary_template.R
\name{copy_summary_template}
\alias{copy_summary_template}
\title{Copy Summary Rmd Template}
\usage{
copy_summary_template(path, name = "summary.Rmd", overwrite = TRUE)
}
\arguments{
\item{path}{The path for the .Rmd and style.css files}

\item{name}{The name for the .Rmd file}
}
\description{
Copy Summary Rmd Template
}
