% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df_B_lofreq}
\alias{example_df_B_lofreq}
\title{Example iVar, lofreq snpSIFT dataframes B}
\format{
## `example_df_B_lofreq`
}
\usage{
example_df_B_lofreq
}
\description{
Example iVar, lofreq snpSIFT dataframes B
}
\seealso{
Other "example datasets": 
\code{\link{example_df_A_ivar}},
\code{\link{example_df_A_lofreq}},
\code{\link{example_df_B_ivar}},
\code{\link{example_df_C_ivar}},
\code{\link{example_df_C_lofreq}}
}
\concept{"example datasets"}
\keyword{datasets}
