% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df_C_ivar}
\alias{example_df_C_ivar}
\title{Example iVar snpSIFT dataframes C}
\format{
## `example_df_C_ivar`
}
\usage{
example_df_C_ivar
}
\description{
Example iVar snpSIFT dataframes C
}
\seealso{
Other "example datasets": 
\code{\link{example_df_A_ivar}},
\code{\link{example_df_A_lofreq}},
\code{\link{example_df_B_ivar}},
\code{\link{example_df_B_lofreq}},
\code{\link{example_df_C_lofreq}}
}
\concept{"example datasets"}
\keyword{datasets}
