% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverages.R
\name{get_coverages}
\alias{get_coverages}
\title{Make coverage files from folder with bedgraphs}
\usage{
get_coverages(file_path)
}
\arguments{
\item{file_path}{A directory with bedgraph files}
}
\description{
Make coverage files from folder with bedgraphs
}
