% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_info.R
\name{get_global_outbreak_info}
\alias{get_global_outbreak_info}
\title{Global data from outbreak API}
\usage{
get_global_outbreak_info(SNVs)
}
\arguments{
\item{SNVs}{A SNV or vector of SNVs formatted for the outbreak API (eg S:Y1155H)}
}
\description{
Global data from outbreak API
}
\examples{
get_global_outbreak_info(SNVs = "S:Y1155H")
}
