% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_info.R
\name{get_outbreak_info}
\alias{get_outbreak_info}
\title{Data from outbreak API}
\usage{
get_outbreak_info(SNVs, states = "California")
}
\arguments{
\item{SNVs}{A SNV or vector of SNVs formatted for the outbreak API (eg S:Y1155H)}
}
\description{
Data from outbreak API
}
\examples{
get_outbreak_info(SNVs = "S:Y1155H")

}
