% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_dates.R
\name{get_outbreak_latest}
\alias{get_outbreak_latest}
\title{Get most recent dates and proportions from outbreak API data}
\usage{
get_outbreak_latest(df)
}
\arguments{
\item{df}{A tibble or dataframe from one of the get outbreak functions}
}
\description{
Percents are determined to 0.001% accuracy using the scales package.
}
\examples{
get_outbreak_latest(top20_results)
get_outbreak_info(SNVs = "S:Y1155H") |> get_outbreak_latest()
}
