% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_dates.R
\name{get_outbreak_peak}
\alias{get_outbreak_peak}
\title{Get peak dates and proportions from outbreak API data}
\usage{
get_outbreak_peak(df)
}
\arguments{
\item{df}{A tibble or dataframe from one of the get outbreak functions}
}
\description{
If multiple dates have the same proportion, only the fist date is shown.
}
\details{
Percents are determined to 0.001% accuracy using the scales package.
}
