% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_dates.R
\name{get_outbreak_start}
\alias{get_outbreak_start}
\title{Get start dates and proportions from outbreak API data}
\usage{
get_outbreak_start(df)
}
\arguments{
\item{df}{A tibble or dataframe from one of the get outbreak functions}
}
\description{
Percents are determined to 0.001% accuracy using the scales package.
}
\examples{
get_outbreak_start(top20_results)
get_outbreak_info(SNVs = "S:Y1155H") \%>\% get_outbreak_start()
}
