% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_info.R
\name{get_state_outbreak_info}
\alias{get_state_outbreak_info}
\title{State data from outbreak API}
\usage{
get_state_outbreak_info(states, SNVs)
}
\arguments{
\item{states}{A state or vector of states}

\item{SNVs}{A SNV or vector of SNVs formatted for the outbreak API (eg S:Y1155H)}
}
\description{
State data from outbreak API
}
\examples{
get_state_outbreak_info(states = c("CA", "OR"), SNVs = "S:Y1155H")
}
