% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{gini_simpson}
\alias{gini_simpson}
\title{Calculate the Gini-Simpson index for a set of frequencies}
\usage{
gini_simpson(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the Gini-Simpson index
}
\description{
Calculate the Gini-Simpson index for a set of frequencies
}
\details{
The Gini-Simpson is 1 minus the probability that two bases are the
         same. Aka 1 minus Simpson's index. Aka 1 minus 1/N_2
}
\seealso{
vegan::diversity
}
