% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_venn.R
\name{make_venn}
\alias{make_venn}
\title{Make Pipeline Venn}
\usage{
make_venn(df, slices = "SNV")
}
\arguments{
\item{df}{Dataframe or tibble}

\item{slices}{Column name to fill the slices of the venn}
}
\description{
Make Pipeline Venn
}
