% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_alt.R
\name{parse_ref_alt}
\alias{parse_ref_alt}
\title{Parse REF and ALT AA Characters}
\usage{
parse_ref_alt(df)
}
\arguments{
\item{df}{A dataframe of imported snpSIFT data from mappgene output}
}
\description{
The HGVS_P is parsed according to 11 different rules, depending on the EFFECT,
and if there is a deletion, insertion or duplication.
}
