% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverages.R
\name{plot_coverage_ecdf}
\alias{plot_coverage_ecdf}
\title{Plot ECDF coverage}
\usage{
plot_coverage_ecdf(coverage, pseudocount = 1, intercept = 100)
}
\arguments{
\item{coverage}{A coverage from dataframe}

\item{pseudocount}{A pseudocount to add to coverage to make log10 transformation work}

\item{intercept}{An x value to draw a vertical line}
}
\description{
Plot ECDF coverage
}
