% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_counts.R
\name{plot_effect_counts}
\alias{plot_effect_counts}
\title{Barchart summary of the SNPsift Effects}
\usage{
plot_effect_counts(df)
}
\arguments{
\item{df}{Dataframe or tibble}
}
\description{
Barchart summary of the SNPsift Effects
}
