% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap of Specific Gene}
\usage{
plot_heatmap(df, columns = "SAMPLE", rows = "SNV", na.value = 0)
}
\arguments{
\item{df}{A tibble or dataframe from one of the get outbreak functions}

\item{columns}{The column name from df to use as columns in the heatmap}

\item{rows}{The column name from df to use as rows in the heatmap}

\item{na.value}{The numerical value to set the NA values to. Changing this (from 0 to -1) can affect the "chaining" effect of clustering.}
}
\description{
If working with a lot of data it is recommended to filter your df first, and then pass it into the plot_heatmap() function
}
\examples{
plot_heatmap(df, rows = "SNV", columns = "SAMPLE")
df |> filter(GENE == "ORF3a") |> plot_heatmap(rows = "EFFECT", columns = "SAMPLE")
}
