% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmds.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Plot NMDS of AF values}
\usage{
plot_nmds(df, points = "SAMPLE", species = "SNV", distance = "bray")
}
\arguments{
\item{df}{Dataframe or tibble}

\item{points}{The main values to ordinate}

\item{species}{The secondary values to order the points with}

\item{distance}{A distance method}
}
\description{
If you don't get convergent, try changing the seed with set.seed()
}
