% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_outbreak_dates.R
\name{plot_outbreak_milestones}
\alias{plot_outbreak_milestones}
\title{Get the first, peak and most recent times and plots for a list of mutations}
\usage{
plot_outbreak_milestones(df, snv_list, ncol = 2, min.date = NA, max.date = NA)
}
\arguments{
\item{df}{A tibble or dataframe from get_outbreak_info}

\item{snv_list}{A list of the SNVs}

\item{ncol}{The number of columns to facet the plots}

\item{min.date}{Filter the df to only include dates later than or equal to this date (YYYY-MM-DD)}

\item{max.date}{Filter the df to only include dates earlier than or equal to this date (YYYY-MM-DD)}
}
\description{
Get the first, peak and most recent times and plots for a list of mutations
}
