% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outbreak_info.R
\name{plot_outbreak_trends}
\alias{plot_outbreak_trends}
\title{Plot trend data by location from outbreak API}
\usage{
plot_outbreak_trends(
  df,
  title = "Proportions (free y-axis scale)",
  line_width = 1,
  line_alpha = 0.8
)
}
\arguments{
\item{df}{A tibble or dataframe from one of the get outbreak functions}

\item{title}{Override the default title}
}
\value{
A ggplot2 object which can be modified inline
}
\description{
Plot trend data by location from outbreak API
}
