% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mappgene_data.R
\name{read_mappgene_data}
\alias{read_mappgene_data}
\title{Load the mappgene data into a tibble}
\usage{
read_mappgene_data(
  path,
  workflow = c("ivar", "lofreq"),
  sample_cleanup = "LLNL_"
)
}
\arguments{
\item{path}{The path to the folder with the *.ivar.lofreq.snpSIFT.txt files.}

\item{workflow}{The mappgene workflow type(s). Accepts a single string, or a vector of both ivar and lofreq.}

\item{sample_cleanup}{Additional text to remove from the sample names.}
}
\description{
Load the mappgene data into a tibble
}
