% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{simpson}
\alias{simpson}
\title{Calculate the Simpson index for a set of frequencies}
\usage{
simpson(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the Simpson index
}
\description{
Calculate the Simpson index for a set of frequencies
}
\details{
The Simpson index is the probability that two bases are the
         same. Aka the reciprocal of N_2. Aka π for 1 nt.
}
\seealso{
vegan::diversity
}
