% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_alt.R
\name{view_HGVS_C}
\alias{view_HGVS_C}
\title{View HGVS_C Effects}
\usage{
view_HGVS_C(HGVS_C, gene = "S", window = 10)
}
\arguments{
\item{HGVS_C}{A string from snpSIFT output}

\item{gene}{The gene name}

\item{window}{An approximate upstream and downstream window size for plotting}
}
\description{
Plot the consequences of HGVS_C against the wildtype (NC_045512.2) gene sequence
}
