% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zika_reference}
\alias{zika_reference}
\title{Zika genome coordinates reference}
\format{
## `zika_reference`
}
\usage{
zika_reference
}
\description{
Zika genome coordinates reference
}
\keyword{datasets}
