% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zika_reference.R
\name{add_zika_reference}
\alias{add_zika_reference}
\title{Add data for processed Zika peptides}
\usage{
add_zika_reference(df)
}
\arguments{
\item{df}{A mappgene summary dataframe}
}
\description{
Add data for processed Zika peptides
}
