% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{hill_12I}
\alias{hill_12I}
\title{Calculate Hill numbers of order 1, 2, and Inf from a matrix}
\usage{
hill_12I(x)
}
\arguments{
\item{x}{a matrix with nrows samples and ncol counts}
}
\value{
a data.frame with Hill's N_1, N_2, N_Inf for each sample
}
\description{
Calculate Hill numbers of order 1, 2, and Inf from a matrix
}
\details{
log(N_1) is entropy, log(N_2) is inverse Simpson's
}
