% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_counts.R
\name{plot_effect_counts}
\alias{plot_effect_counts}
\title{Barchart summary of the SNPsift Effects}
\usage{
plot_effect_counts(
  df,
  keeper_effects = c("stop_gained", "missense_variant",
    "conservative_inframe_deletion", "conservative_inframe_insertion",
    "disruptive_inframe_deletion", "disruptive_inframe_insertion")
)
}
\arguments{
\item{df}{Dataframe or tibble}
}
\description{
Barchart summary of the SNPsift Effects
}
